features/use-context-api-in-the-app - Use context API to avoid the unnecessary props drilling
