+++ src/components/container/header/Header.tsx
@@ -3,17 +3,13 @@ import Logo from '@/components/container/header/Logo';
 import PageHeading from '@/components/container/header/PageHeading';
 import FeedbackForm from '@/components/container/header/FeedbackForm';
 
-type HeaderProps = {
-  handleAddToList: (text: string) => void;
-};
-
-const Header = ({ handleAddToList }: HeaderProps) => {
+const Header = () => {
   return (
     <header>
       <Pattern />
       <Logo />
       <PageHeading />
-      <FeedbackForm handleAddToList={handleAddToList} />
+      <FeedbackForm />
     </header>
   );
 };
