+++ src/main.tsx
@@ -1,12 +1,16 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
 
-import App from '@/App.tsx'
+import App from '@/App.tsx';
 
-import '@/index.css'
+import AppContextProvider from '@/contexts/AppContextProvider';
+
+import '@/index.css';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <App />
+    <AppContextProvider>
+      <App />
+    </AppContextProvider>
   </StrictMode>,
-)
+);
